requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7
sqlparse==0.4.3
